chore: improved websocket/std redirection logic

+++ corerl/web/agent_manager.py
@@ -1,3 +1,8 @@
+"""
+Agent Manager endpoints for starting and stopping our `corerl_main` scripts.
+Currently, this logic will only function as expected using one web worker!
+"""
+
 import asyncio
 import json
 import logging
@@ -5,19 +10,22 @@ import os
 import sqlite3
 import subprocess
 import tempfile
-import threading
+import uuid
 from collections import deque
 from pathlib import Path
-from typing import IO, Dict, List, Optional, TypedDict
+from threading import Thread
+from typing import Dict, List, Literal, Optional, TypedDict
 
 import yaml
+from anyio import ClosedResourceError
 from fastapi import APIRouter, WebSocket, WebSocketDisconnect, status, websockets
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 from corerl.config import MainConfig
 from corerl.configs.errors import ConfigValidationErrors
 from corerl.configs.loader import config_from_dict, config_to_json
+from corerl.utils.time import now_iso
 from corerl.web import get_coreio_sqlite_path
 
 logger = logging.getLogger(__name__)
@@ -26,53 +34,94 @@ logger.setLevel(logging.DEBUG)
 router = APIRouter()
 
 
+class WSMessage(BaseModel):
+    id: uuid.UUID = Field(default_factory=uuid.uuid4)
+    time: str = Field(default_factory=now_iso)
+    type: Literal["log"] = "log"
+    message: str | None = None
+
+
 class CoreRLProcess(TypedDict):
     """TypedDict for CoreRL processes managed within the FastAPI server."""
 
-    process: subprocess.Popen
-    thread: threading.Thread
-    log_buffer: deque[str]
-    temp_config_file: str
+    log_buffer: deque[WSMessage]
+    process: subprocess.Popen | None
 
 
 processes: Dict[str, CoreRLProcess] = {}
 clients: Dict[str, set[WebSocket]] = {}
 
 
-async def read_stream(stream: IO[bytes] | None, config_id: str, temp_config_file: str):
-    """Reads a subprocess stream and send it to connected websocket clients."""
-    coreRLProcess = processes[config_id]
-    process = coreRLProcess["process"]
-    log_buffer = coreRLProcess["log_buffer"]
-
-    while process.poll() is None:
-        line = stream.readline() if stream else None
-        if not line:
-            continue
-        line = line.decode("utf-8").rstrip()
-        logger.debug(f"[{config_id}] {line}")
-        log_buffer.appendleft(line)
-        if config_id in clients:
-            for client in clients[config_id]:
-                try:
-                    await client.send_text(line)
-                except Exception:
-                    logger.exception("Error sending websocket message")
-                    continue
-
-        if os.path.exists(temp_config_file):
-            os.remove(temp_config_file)
-
-    else:
-        # process has terminated
-        if os.path.exists(temp_config_file):
-            os.remove(temp_config_file)
-
-
 class StartAgentRequestPayload(BaseModel):
     """Payload for starting an agent."""
 
     extra_options: Optional[List[str]] = None
+
+
+async def start_proc_read_stream(yaml_config: str, config_id: str, req_payload: StartAgentRequestPayload):
+    """Start our process, send all stdout and stderr to connected websocket clients."""
+    Path("config").mkdir(exist_ok=True)
+
+    # cannot delete the temporary configuration file within the context manager because
+    # there is an unspecified delay between context manager exit and subprocess's python script running
+    # instead, remove the file after some output has been emitted from the process
+    with tempfile.NamedTemporaryFile(suffix=".yaml", dir="config", delete=False) as temp_file:
+        temp_file.write(yaml_config.encode())
+        temp_file.seek(0)
+        logging.info(f"Temporary config file created at: {temp_file.name}")
+
+        cmd = [
+            "corerl_main",
+            "--config-name",
+            temp_file.name,
+        ]
+        if req_payload.extra_options:
+            cmd.extend(req_payload.extra_options)
+
+    unbuffered_env = os.environ.copy()
+    unbuffered_env["PYTHONUNBUFFERED"] = "true"
+    process = subprocess.Popen(
+        cmd,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,  # redirect stderr to stdout
+        universal_newlines=False,
+        start_new_session=True,
+        env=unbuffered_env,
+    )
+
+    processes[config_id]["process"] = process
+    log_buffer = processes[config_id]["log_buffer"]
+
+    stream = process.stdout
+    assert stream is not None
+
+    while process.poll() is None:
+        line = stream.readline()
+        if not line:
+            continue
+
+        if os.path.exists(temp_file.name):
+            os.remove(temp_file.name)
+
+        line = line.decode("utf-8").rstrip()
+        ws_msg = WSMessage(message=line)
+        log_buffer.append(ws_msg)
+
+        logger.debug(f"[{config_id}] {line}")
+
+        if config_id not in clients:
+            continue
+
+        for client in clients[config_id]:
+            if client.application_state != websockets.WebSocketState.CONNECTED:
+                continue
+            try:
+                await client.send_json(json.loads(ws_msg.model_dump_json()))
+            except ClosedResourceError:
+                pass
+            except Exception:
+                logger.exception("Error sending websocket message")
+                continue
 
 
 @router.post("/{config_id}/start")
@@ -82,7 +131,10 @@ def agent_start(req_payload: StartAgentRequestPayload, config_id: str):
     and start the core-rl agent, tracking the process id.
     """
 
+    if config_id in processes:
+        return {"status": "ERROR", "message": "agent already started, must stop before restart"}
     try:
+        processes[config_id] = CoreRLProcess({"log_buffer": deque(maxlen=20000), "process": None})
         # get the raw MainConfig JSON payload from CoreIO's sqlite DB
         with sqlite3.connect(get_coreio_sqlite_path()) as conn:
             cur = conn.cursor()
@@ -106,58 +158,13 @@ def agent_start(req_payload: StartAgentRequestPayload, config_id: str):
         json_config = json.loads(config_to_json(MainConfig, res_config))
         yaml_config = yaml.safe_dump(json_config, sort_keys=False)
 
-        Path("config").mkdir(exist_ok=True)
+        t = Thread(target=asyncio.run, args=(start_proc_read_stream(yaml_config, config_id, req_payload),), daemon=True)
+        t.start()
 
-        # Cannot delete the file within the context manager because subprocess may take some time
-        with tempfile.NamedTemporaryFile(suffix=".yaml", dir="config", delete=False) as temp_file:
-            temp_file.write(yaml_config.encode())
-            temp_file.seek(0)
-            logging.info(f"Temporary config file created at: {temp_file.name}")
-
-            cmd = [
-                "corerl_main",
-                "--config-name",
-                temp_file.name,
-            ]
-            if req_payload.extra_options:
-                cmd.extend(req_payload.extra_options)
-
-            process = subprocess.Popen(
-                cmd,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.STDOUT,  # redirect stderr to stdout
-                universal_newlines=False,
-                start_new_session=True,
-            )
-
-        logger.info(f"Started agent with PID: {process.pid}")
-
-        stdout_thread = threading.Thread(
-            target=asyncio.run,
-            args=(
-                read_stream(
-                    process.stdout,
-                    config_id,
-                    temp_file.name,
-                ),
-            ),
-            daemon=True,
-            name=f"corerl_agent_{config_id}_stdout",
-        )
-        processes[config_id] = {
-            "process": process,
-            "thread": stdout_thread,
-            # used to store the last 100 lines of log output for each agent
-            "log_buffer": deque(maxlen=100),
-            "temp_config_file": temp_file.name,
-        }
-
-        stdout_thread.start()
-
-        return {"status": "SUCCESS", "pid": process.pid}
+        return {"status": "SUCCESS"}
 
     except Exception as e:
-        return {"status": "ERROR", "message": str(e), "sqlite_path": get_coreio_sqlite_path}
+        return {"status": "ERROR", "message": str(e), "sqlite_path": get_coreio_sqlite_path()}
 
 
 @router.post("/{config_id}/stop")
@@ -168,11 +175,9 @@ def agent_stop(config_id: str):
     if config_id in processes:
         coreRLProcess = processes[config_id]
         process = coreRLProcess["process"]
-        thread = coreRLProcess["thread"]
 
-        if process.poll() is None:
+        if process and process.poll() is None:
             process.terminate()
-        thread.join(timeout=5)
         del processes[config_id]
 
         return {"status": "stopped"}
@@ -187,6 +192,8 @@ def agent_status(config_id: str):
     if config_id in processes:
         coreRLProcess = processes[config_id]
         process = coreRLProcess["process"]
+        if not process:
+            return {"status": "starting_up"}
         if process.poll() is None:
             # Process is still running
             return {"status": "running", "pid": process.pid}
@@ -215,9 +222,9 @@ async def websocket_endpoint(websocket: WebSocket, config_id: str, skip_log_buff
     if not skip_log_buffer and config_id in processes:
         coreRLProcess = processes[config_id]
         log_buffer = coreRLProcess["log_buffer"]
-        for line in log_buffer:
+        for ws_msg in log_buffer:
             try:
-                await websocket.send_text(line)
+                await websocket.send_json(json.loads(ws_msg.model_dump_json()))
             except Exception:
                 logger.exception("Error sending websocket message")
                 continue
